Push mpx properties out of the base class
