+++ src/DataService/Media/MediaFile.php
@@ -4,6 +4,7 @@ namespace Lullabot\Mpx\DataService\Media;
 
 use Lullabot\Mpx\DataService\Annotation\DataService;
 use Lullabot\Mpx\DataService\ObjectBase;
+use Psr\Http\Message\UriInterface;
 
 /**
  * Implements a MediaFile object.
@@ -409,6 +410,24 @@ class MediaFile extends ObjectBase
      * @var int
      */
     protected $width;
+    /**
+     * The date and time that this object was created.
+     *
+     * @var \DateTime
+     */
+    protected $added;
+    /**
+     * The id of the user that created this object.
+     *
+     * @var \Psr\Http\Message\UriInterface
+     */
+    protected $addedByUserId;
+    /**
+     * The id of the account that owns this object.
+     *
+     * @var \Psr\Http\Message\UriInterface
+     */
+    protected $ownerId;
 
     /**
      * Returns whether playback is enabled for this file, and whether it can be associated with new Release objects.
@@ -1508,5 +1527,69 @@ class MediaFile extends ObjectBase
     public function setWidth($width)
     {
         $this->width = $width;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getAdded(): \DateTime
+    {
+        return $this->added;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setAdded($added)
+    {
+        $this->added = $added;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getAddedByUserId(): \Psr\Http\Message\UriInterface
+    {
+        return $this->addedByUserId;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setAddedByUserId($addedByUserId)
+    {
+        $this->addedByUserId = $addedByUserId;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getId(): \Psr\Http\Message\UriInterface
+    {
+        return $this->id;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setId(UriInterface $id)
+    {
+        $this->id = $id;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getOwnerId(): \Psr\Http\Message\UriInterface
+    {
+        return $this->ownerId;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setOwnerId($ownerId)
+    {
+        $this->ownerId = $ownerId;
     }
 }
